+++ packages/perseus/src/__tests__/widget-container.new.test.tsx
@@ -6,27 +6,28 @@
 // TODO(LEMS-4304): feature flag cleanup - rename this file to widget-container.test.tsx.
 // This file is the new widget container test file that will replace the old container tests.
 
+import {linterContextDefault} from "@khanacademy/perseus-linter";
 import {render, screen} from "@testing-library/react";
 import * as React from "react";
 
 import * as Dependencies from "../dependencies";
+import {ApiOptions} from "../perseus-api";
 import {
     testDependencies,
     testDependenciesV2,
 } from "../testing/test-dependencies";
+import {containerSizeClass} from "../util/sizing-utils";
 import WidgetContainer from "../widget-container.new";
 import {registerWidget} from "../widgets";
 import Explanation from "../widgets/explanation";
 import Image from "../widgets/image";
 
-import type {PerseusDependenciesV2, WidgetExports} from "../types";
+import type {PerseusDependenciesV2, WidgetExports, WidgetProps} from "../types";
 
 const MockWidgetComponent = ({
-    text,
-    fail = false,
+    options: {text, fail = false},
 }: {
-    text: string;
-    fail: boolean;
+    options: {text: string; fail: boolean};
 }) => {
     if (fail) {
         throw new Error("MockWidget failed to render");
@@ -41,6 +42,30 @@ const MockWidget: WidgetExports<typeof MockWidgetComponent> = {
     widget: MockWidgetComponent,
 };
 
+const getBaseProps = (
+    // TODO(benchristel): Pass real type arguments here. Maybe getBaseProps can
+    //  be generic.
+    overrides?: Partial<WidgetProps<any, any>>,
+): WidgetProps<any, any> => ({
+    options: {},
+    trackInteraction: () => {},
+    widgetId: "widget 1",
+    widgetIndex: 0,
+    alignment: null,
+    static: false,
+    problemNum: 0,
+    apiOptions: {...ApiOptions.defaults, isMobile: false},
+    onFocus: () => {},
+    onBlur: () => {},
+    findWidgets: () => [],
+    reviewMode: false,
+    handleUserInput: () => {},
+    userInput: {},
+    linterContext: linterContextDefault,
+    containerSizeClass: containerSizeClass.MEDIUM,
+    ...overrides,
+});
+
 describe("widget-container", () => {
     it("should render nothing when requested widget not registered", () => {
         // Arrange
@@ -51,7 +76,7 @@ describe("widget-container", () => {
             <WidgetContainer
                 type="invalid-widget"
                 id="invalid-widget 1"
-                widgetProps={{apiOptions: {isMobile: false}}}
+                widgetProps={getBaseProps()}
             />,
         );
 
@@ -77,16 +102,14 @@ describe("widget-container", () => {
                 <WidgetContainer
                     type="explanation"
                     id="explanation 1"
-                    widgetProps={{
-                        showPrompt: "Explanation",
-                        hidePrompt: "Hide explanation",
-                        explanation: "This is an explanation",
-                        widgets: {},
-
-                        findWidgets: () => [],
-
-                        apiOptions: {isMobile: false},
-                    }}
+                    widgetProps={getBaseProps({
+                        options: {
+                            showPrompt: "Explanation",
+                            hidePrompt: "Hide explanation",
+                            explanation: "This is an explanation",
+                            widgets: {},
+                        },
+                    })}
                 />
             </Dependencies.DependenciesContext.Provider>,
         );
@@ -107,7 +130,7 @@ describe("widget-container", () => {
 
         const renderContainer = (
             type: string,
-            widgetProps: Record<string, unknown>,
+            widgetProps: Partial<WidgetProps<any, any>>,
         ) => {
             const {container} = render(
                 <Dependencies.DependenciesContext.Provider
@@ -116,11 +139,7 @@ describe("widget-container", () => {
                     <WidgetContainer
                         type={type}
                         id={`${type} 1`}
-                        widgetProps={{
-                            findWidgets: () => [],
-                            apiOptions: {isMobile: false},
-                            ...widgetProps,
-                        }}
+                        widgetProps={getBaseProps(widgetProps)}
                     />
                 </Dependencies.DependenciesContext.Provider>,
             );
@@ -132,8 +151,7 @@ describe("widget-container", () => {
         it("renders a <div> for a widget with no alignment specified", () => {
             // Arrange, Act
             const containerElement = renderContainer("mock-widget", {
-                text: "Hello world!",
-                fail: false,
+                options: {text: "Hello world!", fail: false},
             });
 
             // Assert
@@ -143,7 +161,7 @@ describe("widget-container", () => {
         it("renders a <div> for a block-aligned widget", () => {
             // Arrange, Act - the image widget's default alignment is "block".
             const containerElement = renderContainer("image", {
-                backgroundImage: {},
+                options: {backgroundImage: {}},
             });
 
             // Assert
@@ -153,10 +171,12 @@ describe("widget-container", () => {
         it("renders a <span> for an inline-aligned widget", () => {
             // Arrange, Act
             const containerElement = renderContainer("explanation", {
-                showPrompt: "Explanation",
-                hidePrompt: "Hide explanation",
-                explanation: "This is an explanation",
-                widgets: {},
+                options: {
+                    showPrompt: "Explanation",
+                    hidePrompt: "Hide explanation",
+                    explanation: "This is an explanation",
+                    widgets: {},
+                },
             });
 
             // Assert
@@ -167,10 +187,12 @@ describe("widget-container", () => {
             // Arrange, Act
             const containerElement = renderContainer("explanation", {
                 alignment: "default",
-                showPrompt: "Explanation",
-                hidePrompt: "Hide explanation",
-                explanation: "This is an explanation",
-                widgets: {},
+                options: {
+                    showPrompt: "Explanation",
+                    hidePrompt: "Hide explanation",
+                    explanation: "This is an explanation",
+                    widgets: {},
+                },
             });
 
             // Assert
@@ -181,10 +203,12 @@ describe("widget-container", () => {
             // Arrange, Act
             const containerElement = renderContainer("explanation", {
                 alignment: "block",
-                showPrompt: "Explanation",
-                hidePrompt: "Hide explanation",
-                explanation: "This is an explanation",
-                widgets: {},
+                options: {
+                    showPrompt: "Explanation",
+                    hidePrompt: "Hide explanation",
+                    explanation: "This is an explanation",
+                    widgets: {},
+                },
             });
 
             // Assert
@@ -216,14 +240,9 @@ describe("widget-container", () => {
                 <WidgetContainer
                     type="mock-widget"
                     id="mock-widget 1"
-                    widgetProps={{
-                        text: "Hello world!",
-                        fail: true,
-
-                        findWidgets: () => [],
-
-                        apiOptions: {isMobile: false},
-                    }}
+                    widgetProps={getBaseProps({
+                        options: {text: "Hello world!", fail: true},
+                    })}
                 />
             </Dependencies.DependenciesContext.Provider>,
         );